gmail-insights-ai-agent/pyproject.toml
[project]
name = "gmail-insights-ai-agent"
version = "0.1.0"
description = "AI Agent that connects to Gmail, stores key partner communication insights in GibsonAI database, and helps users recall conversations with context-aware memory"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=1.7.0",
    "anthropic>=0.55.0",
    "python-dotenv>=1.1.1",
    "groq>=0.20.0",
    "mcp>=1.10.0",
    "openai>=1.70.0",
    "streamlit>=1.44.0",
    "pydantic>=2.11.7",
    "python-dateutil>=2.9.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py310"
line-length = 88
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".env",
    "tmp",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
